Adds LaTeX package builder for Font Awesome 6
Implements a builder to generate a LaTeX package from Font Awesome 6 metadata and font files.

This commit introduces a new `TexBuilder` class that automates the process of creating a `.sty` file containing LaTeX commands for each Font Awesome icon.

It also includes functionality to copy the necessary font files and licenses to the output directory.